requirements.txt
openai==0.27.0
python-dotenv==0.19.2
pytest==7.1.2
black==22.3.0
flake8==4.0.1
